% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/API_Auth.R
\name{auth_key}
\alias{auth_key}
\title{Assign Elsevier API tokens}
\usage{
auth_key(x)
}
\arguments{
\item{x}{- an empty parameter, e.g. NULL}
}
\description{
There are two API Keys. One is for the access of almost everything, except for text-mining
capabilities (Elsevier_API). Another one (Elsevier_TM_API) is only for the purpose of using text-mining
functionality. See URL for your restrictions below. The first one is always required whereas the
second one is optional.
}
\examples{
options(Elsevier_API = "F1QH-Q64B-BSBI-JASJ", Elsevier_TM_API = "1QH-Q64B-BSBI-JAS")
}
\seealso{
Register your Keys \url{http://dev.elsevier.com/myapikey.html}

See for restrictions which apply to your key \url{http://dev.elsevier.com/api_key_settings.html}

See \url{https://github.com/ropensci/rnoaa/blob/master/R/zzz.r#L145}

See \url{https://github.com/ropensci/gistr/blob/master/R/gist_auth.R#L24}
}

