% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Search.R
\name{Citations_Overview_Json}
\alias{Citations_Overview_Json}
\title{Test with JSON}
\usage{
Citations_Overview_Json(query)
}
\arguments{
\item{ScienceDirect}{search query string}
}
\description{
Test with JSON
}

