% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Citated.R
\name{citatissson_count}
\alias{citatissson_count}
\title{Cited by in Scopus}
\usage{
citatissson_count(apiKey = auth_key(NULL), showFull = TRUE)
}
\description{
With Scopus Abstract Citation Count API you want to show the
number of times the papers have been cited by other papers covered by Scopus.
}
\examples{
Citation_Count(doi ="10.1016/S0014-5793(01)03313-0")
}
\seealso{
\url{http://dev.elsevier.com/tecdoc_cited_by_in_scopus.html}

\url{http://api.elsevier.com/documentation/AbstractCitationCountAPI.wadl}
}

