% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Search.R
\name{ScienceDirect_SearchJson}
\alias{ScienceDirect_SearchJson}
\title{Test with JSON}
\usage{
ScienceDirect_SearchJson(query)
}
\arguments{
\item{ScienceDirect}{search query string}
}
\description{
Test with JSON
}
\seealso{
See \url{http://dev.elsevier.com/api_key_settings.html}
http://dev.elsevier.com/tecdoc_api_authentication.html
Load libraries & execute code manually

library(httr)
library(jsonlite)
}

