% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Search.R
\name{ScienceDirect_SearchXML}
\alias{ScienceDirect_SearchXML}
\title{Test with XML}
\usage{
ScienceDirect_SearchXML(query)
}
\arguments{
\item{ScienceDirect}{search query string

library(xml2)
library(XML)}
}
\description{
Test with XML
}

