% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/textMining.R
\name{textM}
\alias{textM}
\title{Text Mining methods}
\usage{
textM(isnn, apiKey = auth_key(NULL), showRequestURL = F)
}
\description{

}
\details{


c2494d00460743fb8a299ca3de737083
}
\examples{
textM()
}

