% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Api.R
\name{elsevierApi}
\alias{elsevierApi}
\title{Assign Elsevier API token}
\usage{
elsevierApi(auth_token)
}
\arguments{
\item{auth_token}{- Passed parameter to set Bit.ly Generic Access Token \code{\link{elsevierApi}}.}
}
\description{
Assign Elsevier API token
}
\examples{
elsevierApi("7f5eff1eb026_dDOG6dff5574fc165e297ee") ## invalid key
}
\seealso{
See \url{http://dev.elsevier.com/myapikey.html}
}

