% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Search.R
\name{ScienceDirect_SearchJson}
\alias{ScienceDirect_SearchJson}
\title{Test with JSON}
\usage{
ScienceDirect_SearchJson(query)
}
\arguments{
\item{ScienceDirect}{search query string

Load libraries & execute code manually

library(httr)
library(jsonlite)}
}
\description{
Test with JSON
}

